Using the cool -C flag for paths
Entering the command `a ; b ; c` results in the exit code of the last command, `c`, being recorded. So when we're running the Make test things for each of the projects, we're cd-ing into the directory, running the tests, and then cd-ing back to top... and the exit code of that last cd (which is always 0) is what's returned! I'm hypothesizing that this is the issue.